account-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
python-dotenv==0.19.0
PyJWT==2.8.0